+++ lib/rust/zephyr/src/drivers.rs
@@ -1,0 +1,1 @@
+pub mod gpio;
